do.nothing {tao}	R Documentation

If nothing is done, then all will be well.

Description

Nothing is done, nothing is left undone.

Usage

do.nothing(...)

do.nothing.else(...)

Arguments

...
Arguments avail not.

Value

What difference between yes and no?
What difference between success and failure?
Must you value what others value, avoid what others avoid?
How ridiculous!

Author(s)

Lao Tsu, Aaron A. King

Examples

do.nothing()

## Not run: 
  do.nothing.else()

## End(Not run)

[Package tao version 1.1.1 Index]