tao-package {tao}	R Documentation

That which cannot be packaged.

Description

The package that can be described is not the great Tao package.

Details

The package that can be documented is not the eternal package
The procedure that terminates is not the eternal procedure
The undocumented is the origin of feature and bug
The documentation is the mother of limitless confusion
Thus, constantly without desire, one observes the runtime errors
Constantly with desire, one observes the progress bar
These two emerge together but differ in name
Their unity is said to be a mystery
Mystery of mysteries, the door to the next revision

Author(s)

Lao Tsu, Aaron A. King

References

Lao Tsu, Tao Te Ching.

See Also

do.nothing, do.nothing.else

[Package tao version 1.1.1 Index]